+++ packages/app/src/game/store/game.selectors.ts
@@ -25,7 +25,14 @@ export const gameHistoryBestTimeSelector = createSelector(gameSelector, state =>
 );
 export const gameHistoryDifficultySelector = (difficulty: DifficultyEnum) =>
     createSelector(gameSelector, state => state.historyByDifficulty[difficulty]);
+export const gameCompletedGamesSelector = (difficulty: DifficultyEnum) =>
+    createSelector(gameSelector, state => state.historyByDifficulty[difficulty].completedGames);
+export const gameCompletedGameByIdSelector = (difficulty: DifficultyEnum, completedAt: number) =>
+    createSelector(gameSelector, state =>
+        state.historyByDifficulty[difficulty].completedGames.find(game => game.completedAt === completedAt)
+    );
 export const gameIsStartedSelector = createSelector(gameSelector, state => isNotEmptyString(state.sudokuString));
-export const gameIsChallengeModeSelector = createSelector(gameSelector, state => state.isChallengeMode);
-export const gameOpponentStepsSelector = createSelector(gameSelector, state => state.opponentSteps);
-export const gameOpponentTotalTimeSelector = createSelector(gameSelector, state => state.opponentTotalTime);
+export const gameIsChallengeModeSelector = createSelector(gameSelector, state => isNotEmptyString(state.challengeState));
+export const gameChallengeStepsSelector = createSelector(gameSelector, state => state.challengeSteps);
+export const gameChallengeTimeSelector = createSelector(gameSelector, state => state.challengeTime);
+export const gameSolutionsStepsSelector = createSelector(gameSelector, state => state.solutionSteps);