+++ packages/encoder/src/index.ts
@@ -1,0 +1,5 @@
+export type { SolutionStepInterface } from './interfaces/solution-step.interface';
+
+export { Solution } from './classes/solution/solution';
+export { SudokuStringEncoder } from './classes/sudoku-string-encoder/sudoku-string-encoder';
+export { GameStateSerializer } from './classes/game-state-serializer/game-state-serializer';