+++ packages/encoder/src/util/string-to-uint8array.util.ts
@@ -1,0 +1,8 @@
+export const stringToUint8Array = (binary: string): Uint8Array => {
+    const bytes = new Uint8Array(binary.length);
+    for (let i = 0; i < binary.length; i += 1) {
+        bytes[i] = binary.charCodeAt(i);
+    }
+
+    return bytes;
+};